cleanup and refactor in main package
