--- src/main/java/pl/coderslab/main/MainService.java
@@ -1,13 +1,0 @@
-package pl.coderslab.main;
-
-import pl.coderslab.solution.*;
-
-public class MainService {
-
-    private final static SolutionDao SOLUTION_DAO = new SolutionDao();
-
-    public static SolutionDto[] getRecentSolutions(int limit) {
-        SolutionEntity[] entities = SOLUTION_DAO.findRecentSolution(limit);
-        return SolutionMapper.mapSolutionToDtoArray(SolutionMapper.mapEntityToSolutionArray(entities));
-    }
-}
