+++ src/main/java/pl/coderslab/main/ManageExerciseController.java
@@ -1,7 +1,7 @@
 package pl.coderslab.main;
 
 import pl.coderslab.exercise.ExerciseDto;
-import pl.coderslab.userGroup.UserGroupDto;
+import pl.coderslab.exercise.ExerciseService;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
